fix: mobile header styling issues
- Removed redundant unused file.
